update app component
