+++ src/components/app/index.ts
@@ -1,0 +1,2 @@
+export { App } from './App';
+export type { AppStateType } from './App';
