--- src/styles/index.module.scss
@@ -1,4 +1,0 @@
-.main {
-	--bg-color: #ffffff;
-	background-color: var(--bg-color, #ffffff);
-}
